requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
Flask==1.1.2
flask-request-validator==4.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
six==1.15.0
Werkzeug==1.0.1
